Update README.md to include entire documentation
I wrote everything I could think of, I included all sections specified in the rubric and requirements for the doc.  Feel free to add on or change anything you want. Anywhere that says [Placeholder ...] you guys can write what you see fit because that's not my field of expertise with this project.